+++ src/world.rs
@@ -1,7 +1,7 @@
 use crate::{hittable::Hittable, sphere::Sphere};
 
 impl Hittable for Vec<Sphere> {
-    fn hit<'a>(&'a self, ray: &crate::ray::Ray, t_min: f32, t_max: f32) -> Option<crate::hittable::HitRecord<'a>> {
+    fn hit<'a>(&'a self, ray: &crate::ray::Ray, t_min: f64, t_max: f64) -> Option<crate::hittable::HitRecord<'a>> {
         let mut hit = None;
         let mut closest_so_far = t_max;
